requirements.txt
streamlit==1.31.0
pandas==2.1.0
plotly==5.18.0
openpyxl==3.1.2
xlsxwriter==3.1.2
passlib==1.7.4
python-dotenv==1.0.0
pillow==10.0.1
